Revert "式と演算子の勉強"
