--- Practice4.java
@@ -1,25 +1,0 @@
-class Practice4
-{
-  public static void main(String[] args)
-  {
-    int num = 0 - 4;
-
-    System.out.println("0-4は" + num + "です。");
-
-    double num2 = 3.14 * 2;
-
-    System.out.println("3.14*2は" + num2 + "です。");
-
-    double num3 = (double)5/3;
-
-    System.out.println("5÷3は" + num3 + "です。");
-
-    int num4 = 30%7;
-
-    System.out.println("30÷7の余りの数は" + num4 + "です。");
-
-    double num5 = (double)(7+32)/5;
-
-    System.out.println("(7+32)÷5は" + num5 + "です。");
-  }
-}
