+++ Sample1.java
@@ -2,7 +2,8 @@ class Sample1
 {
   public static void main(String[] args)
   {
-    System.out.println("1+2は" + (1+2) + "です。");
-    System.out.println("3*4は" + (3*4) + "です。");
+    int num;
+    num = 3;
+    System.out.println("変数numの値は" + num + "です。");
   }
 }
