--- Sample10.java
@@ -1,11 +1,0 @@
-class Sample10
-{
-  public static void main(String[] args)
-  {
-    int d = 2;
-    double pi = 3.14;
-
-    System.out.println("直径が" + d + "センチの円の");
-    System.out.println("円周は" + (d*pi) + "センチです。");
-  }
-}
