+++ Sample2.java
@@ -2,16 +2,11 @@ class Sample2
 {
   public static void main(String[] args)
   {
-    int num1 = 2;
-    int num2 = 3;
-    int sum = num1 + num2;
-
-    System.out.println("変数num1の値は" + num1 + "です。");
-    System.out.println("変数num2の値は" + num2 + "です。");
-    System.out.println("num1+num2の値は" + sum + "です。");
-
-    num1 = num1+1;
-
-    System.out.println("変数num1の値に1を足すと" + num1 + "です。");
+    int num;
+    num = 3;
+    System.out.println("変数numの値は" + num + "です。");
+    num = 5;
+    System.out.println("変数numの値を変更しました。");
+    System.out.println("変数numの新しい値は" + num + "です。");
   }
 }
