--- Sample8.java
@@ -1,14 +1,0 @@
-class Sample8
-{
-  public static void main(String[] args)
-  {
-    int inum = 160;
-
-    System.out.println("身長は" + inum + "センチです。");
-
-    System.out.println("double型の変数を代入します。");
-    double dnum = inum;
-
-    System.out.println("身長は" + dnum + "センチです。");
-  }
-}
